% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_gen.R
\name{random_gen}
\alias{random_gen}
\title{Distribution Based Random Number Generator}
\usage{
random_gen(
  x,
  n = 1,
  seed = FALSE,
  seed_n = 1,
  mean = 10,
  sd = 2,
  df = 2,
  df1 = 1,
  df2 = 10,
  size = 100,
  prob = 0.5
)
}
\arguments{
\item{x}{Name of distribution type ("normal", "t", "F", "binomial")}

\item{n}{Number of observations}

\item{seed}{Logical operator, TRUE if setting seed (default to FALSE)}

\item{seed_n}{number for set seed (default to 1)}

\item{mean}{For normal distribution, mean value (default to 10)}

\item{sd}{For normal distribution, standard deviation value (default to 2)}

\item{df}{For t distribution, degrees of freedom (default to 2)}

\item{df1}{For F distribution, numerator degrees of freedom (default to 1)}

\item{df2}{For F distribution, denominator degrees of freedom (default to 10)}

\item{size}{For binomial distribution, number of trials (default to 100)}

\item{prob}{For binomial distribution, probability of success of each trial (default to 0.5)}
}
\value{
a vector of random numbers
}
\description{
Distribution Based Random Number Generator
}
\examples{
random_gen(x = "binomial", n = 5, size = 501, prob = 0.3, seed = TRUE, seed_n = 1)
}
