% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_plot.R
\name{random_plot}
\alias{random_plot}
\title{Plot Generator for Random Distributions}
\usage{
random_plot(
  x,
  n = 1,
  seed = FALSE,
  seed_n = 1,
  mean = 10,
  sd = 2,
  df = 2,
  df1 = 1,
  df2 = 10,
  size = 100,
  prob = 0.5,
  bins = 30
)
}
\arguments{
\item{x}{Name of distribution type}

\item{n}{Number of observations}

\item{seed}{Logical operator, TRUE if setting seed (default to FALSE)}

\item{seed_n}{number for set seed (default to 1)}

\item{mean}{For normal distribution, mean value (default to 10)}

\item{sd}{For normal distribution, standard deviation value (default to 2)}

\item{df}{For t distribution, degrees of freedom (default to 2)}

\item{df1}{For F distribution, numerator degrees of freedom (default to 1)}

\item{df2}{For F distribution, denominator degrees of freedom (default to 10)}

\item{size}{For binomial distribution, number of trials (default to 100)}

\item{prob}{For binomial distribution, probability of success of each trial (default to 0.5)}

\item{bins}{binwidth of graph (default to 30)}
}
\value{
a plot from randomly generated numbers
}
\description{
Plot Generator for Random Distributions
}
\examples{
random_plot("normal", n = 1000, mean = 10, sd = 1, bins = 20)
}
